Refactor arm quantization
* Move quantization files to own folder
* Breakout annotators into one file/ annotator in new folder
	quantizer/quantization_annotation
* Make QuantizationConfig not optional in annotators
	There is a None-check before calling them.
* Move all Quantization config logic into one class
* Remove unsupported qat functionality
* Remove unsupported dynamic annotation
* Remove unused code
* Add docstrings
