+++ backends/arm/quantizer/__init__.py
@@ -1,0 +1,4 @@
+# Copyright 2024 Arm Limited and/or its affiliates.
+#
+# This source code is licensed under the BSD-style license found in the
+# LICENSE file in the root directory of this source tree.
