+++ backends/arm/tosa_quant_utils.py
@@ -33,7 +33,7 @@ def is_quant_node(node: torch.fx.Node):
     # For Rank > 2 Linear layers, the quant node is after the view_copy
     if (
         node.target == exir_ops.edge.aten.addmm.default
-        and list(node.users)[0].target == exir_ops.edge.aten.view_copy.default
+        and consumer_node.target == exir_ops.edge.aten.view_copy.default
     ):
         consumer_consumer_node = list(consumer_node.users)[0]
         return True if consumer_consumer_node.target == q_op else False
